+++ calculator/src/App.css
@@ -1,4 +1,4 @@
-.app {
+body {
   height: 100vh;
   display: flex;
   justify-content: center;
